scraper/requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
pandas==2.2.3
SQLAlchemy==2.0.35
psycopg2-binary==2.9.10
APScheduler==3.10.4
python-dotenv==1.0.1
lxml==5.3.0
playwright==1.47.0
